requirements.txt
# Modelo de IA Local
ollama-python==0.1.7
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0

# Reconhecimento de Voz
SpeechRecognition>=3.10.0
pyaudio>=0.2.11
whisper-openai>=20231117

# Síntese de Voz
pyttsx3>=2.90
gTTS>=2.3.0
pygame>=2.5.0

# Banco de Dados
sqlite3  # Já incluído no Python
sqlalchemy>=2.0.0

# Processamento de Texto
nltk>=3.8
spacy>=3.6.0

# Utilitários
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
aiofiles>=23.1.0
asyncio-throttle>=1.0.2

# Audio e Multimídia
librosa>=0.10.0
soundfile>=0.12.0
webrtcvad>=2.0.10

# Interface (opcional para futuras expansões)
tkinter  # Já incluído no Python
customtkinter>=5.2.0

# Logging e Monitoramento
colorlog>=6.7.0
psutil>=5.9.0